create new file for all robot movements
